files/requirements.txt
librosa==0.10.1
numpy==1.24.3
textblob==0.17.1
supabase==2.3.1
python-dotenv==1.0.0
asyncio==3.4.3
pathlib==1.0.1 